[qt theme] Use GetRect() not GetContentRect() in NativeThemeQt
Neither one is correct, but GetRect() happens to work since it's inlined and just
returns a local member.  GetContentRect isn't available since it's a function
defined in gklayout, and widget does not link to layout.  (It happens to work in
libxul builds due to the way libxul is linked, but it shouldn't be used.)

UltraBlame original commit: 6d568c5c574d878531ae9d079bc1e6644dbf5a72