Bug 466586 - Only expire cache elements that don't have any proxies/observers.  r=stuart sr=vlad a=blocking1.9.1+
The image cache is now composed of two things: a heap of tracked items that
can be expired and evicted, and a superset of that which hashes from URI to
cache element. Cache elements start out in the second set, and are moved to
the first (and start to be tracked for expiry) once they have no observers.

UltraBlame original commit: af9270b650e6164c087fc5b04eab1a47d0b320f2