Bug 560370 - PPC cmov handling logic incorrect (r+edwsmith)
On PPC the cmov logic is generating incorrect code due to
instruction and register lifetimes are not being correctly
handled.

This patch mirrors the code used on the x86 which was fixed
a while back to address this issue.  See bug 535705

UltraBlame original commit: 20583089b31b86c32c9487d1187b876d6343dcee