Bug 572798 - add LIR_callv (r=nnethercote)
Adds LIR_callv for calls to helper functions that return void.

Added a ValidateWriter check that LIR_callv to be paired with ARGTYPE_V,
plus checks for the other obvious pairings, plus a check that callv must
not call a _pure=1 function.

getCallOpcode() returns LIR_callv for ARGTYPE_V, as expected.  This means that
some calls will return LTy_V from LIns::retType(), as expected, but unlike
before.  This in turn can cause a ValidateWriter error if an instruction uses
the result of a void call.  (after all, that's the point).

Each backend was modified to not assign a register or save the result of a void
call.

UltraBlame original commit: df05d03542f711f92a9ea2faab04cd2b3f2af82c