Bug 611189 - Parchment IF Reader Misbehaves; r=bzbarsky a=blocking-final+
This patch makes sure that we don't try to read the value from the editor if we haven't started to initialize the editor yet.
This ensures that the UpdateValueDisplay calls before we start to initialize the editor will always end up reading the value from the content node, and do not rely on the data returned by the editor, which might be stale.

UltraBlame original commit: a664aa828e62d6a5880308ff2190963d0e8f8916