Bug 614832. Refactor EGL surface creation. r=vlad,a=blocking
This factors CreateSurfaceForWindow() out of CreateForWindow() so that we can
reuse it in RenewSurface(). For this to work we also need to factor out
CreateConfig(). The QT implementation is also pulled out because it didn't
share any code.

UltraBlame original commit: 4b233c7bff2c901ee97dd8b15021e3671722efa1