Fix for bug 633934.  r=bzbarsky
Two files need to be patched:

/dom/src/json/nsJSON.cpp
/dom/src/json/test/unit/test_encode.js

The first problem is a wrong nsnull initialization parameter in one of
the nsJSONWriter constructor.

The second problem is caused by related bug 410005 and only happens on
windows and can be solved by calling the clone() method before calling
filesize attribute on a file that has been created and later on modified
by adding extra content.

The BOM check set to true causes this behaviour

 // check BOMs
  // the clone() calls are there to work around -- bug 410005
  var f = writeToFile({},"UTF-8", true).clone();

As discussed Linux is not affected.

The overall 410005 seems to be related to a wrong logic in the mDirty and
MakeDirty() flag in nsLocalFileWin.cpp implementation.

UltraBlame original commit: 5e74b6d598e762b661bbb1cb5841f590cdbea098