Bug 476733 - WinCE Dynamic CAB INF File Production Needed - ugly patch to fix buildbot issues
When landing this patch, the command for producing a ZIP and CAB for fennec
changed from "make package" to "make installer"

Unfortunately, the buildbots know nothing of this change.

This ugly patch makes the "make package" command build the needed xulrunner and fennec ZIPs,
as well as the desired fennec CAB.  With this patch (attachment 641741) we produce the
XULrunner ZIP 3 times, the fennec ZIP 2 times, and the CAB 2 times.

UltraBlame original commit: fad92baa720d637c4d0d0670b6d071fb76a70624