Bug 664996 Fix missing redraw when resuming application r=dougt
Because I'd altered the way surface creation works, it was possible to
have a null buffer in surfaceChanged while having a valid surface size.

Instead of looking at the buffer pointers, I've replaced it with a
boolean that tracks whether the surface size is valid or not.

In addition, there was a typo that meant that synchronised redraws were
being skipped, as the object from mSyncDraws was taken and immediately
discarded.

UltraBlame original commit: 38f69296b20c922c919d8d932e7c4e82ecc24a18