Bug 403524:  Draw correct colors for text-decorations on default-color shadows in quirks mode. r=dbaron
Make the quirks mode text-decoration + text-shadow code draw the
decorations for shadows that do not have a specified color using
the same color used for the un-shadowed decorations, which matches
standards mode behavior.  (The color of unspecified-color shadows
is explicitly undefined in the specification.)

This code will (in a later patch on this bug) be used for both
quirks and standards modes.

UltraBlame original commit: db9466903986bcf2c3aa42a8b9112661a4e8be95