Bug 676315 - ANGLE build problem: d3dx9.h not found unless $DXSDK_DIR points to it - r=joe
In configure, we detect the DXSDK using the windows registry and set MOZ_DIRECTX_SDK_PATH accordingly. This patch fixes a bug whereby we weren't using that to actually find headers, and were relying on the half-universal DXSDK_DIR variable instead, which is not always defined and in case of multiple versions installed may not point to the most recent one.

UltraBlame original commit: 9c43ba0259fa1536a8caf18efeb1995a5555ff85