Bug 674647 - Update OS X mozconfig snippets. r=ted.mielczarek.
Pass --target to configure in mozconfig.leopard. It is needed by the gcc
in 10.6. Without this gcc would look for the 10.6 CRT files in the 10.5 SDK
and fail.

Don't include the kernel version in  universal/mozconfig.common for consistency.

UltraBlame original commit: 7369baed29efd8bd50c63e4d06b876813864e314