Bug 700720 - Fix crash in Favicons.java on Honeycomb [r=lucasr]
On Honeycomb the database schema changes mean that
a query to update the favicon DB fails with a crash.
Modifying the query so that it also updates the URL
field with the existing URL field value (i.e. a no-op
update) fixes the crash.

UltraBlame original commit: 8515bd5b8993133b345a10a5e63845c5832733c9