Bug 704690 - Abort flings on gecko viewport update [r=Cwiiis]
If the Java code is in the middle of a fling when Gecko sends
a viewport update, abort the fling, cancel the velocity, and
re-fling. The re-fling is purely to get the page snapped to an
edge if needed.

UltraBlame original commit: cba7acb26616d17caf05e4d05314754b9f2d6739