Bug 720663 - devicemanagerADB pushDir fallback not always tried. r=gbrown
The devicemanagerADB.py pushDir method has two paths, one that uses zip/unzip
and the other that just pushes each file one at a time. The idea is that the
first path is chosen, and if it fails, the second is tried before bailing out,
but it's possible for the first failure not to be detected in such a way that
activates the fallback path.

Add a try/except block to catch exceptions in the first path and fall back to
the second.

UltraBlame original commit: 63db726e7634fd9ead1f636e5068d6467a88a8bc