Bug 727939. Avoid double painting after invalidation. r=bgirard
This reduces the paint rate on things like runfield
so that we only paint at the rate that we're invalidating.

We do so by not requesting render in endTransaction

It looks like endTransaction is only called by endDrawing
on the gecko thread and only on VirtualLayers. We don't
need the extra requestRender call because we've already
done the rendering.

UltraBlame original commit: ed2f63d8e4b608364ef7bb788f20a2712313eaf6