Bug 748267 part 1.  Refactor js-to-native conversion to make it not assume it's only happening on stack arguments.  r=peterv
The changes in CastableObjectUnwrapper and CallbackObjectUnwrapper are
basically reindentation.  Similar for the actual conversion changes for all the
non-interface types, though for those the actual declaration of the thing being
converted to has moved out, since we may not be converting onto the stack.

UltraBlame original commit: 7ab9330f8337c6b3fbf25ecd571d136f40486a76