Bug 757400: When we change a compartment's debug mode, ensure we throw away all the outdated analyses. r=billm
Always do a cleansing GC when we turn debug mode off, as well as when we
turn it on. If we don't, it's possible to have stale analyses (generated in
the wrong debug mode state) stick around, and we'll assert in JM.

Have ShouldCleanUpEverything recognize gcreason::DEBUG_MODE_GC as a reason
to discard all jit code and analyses.

UltraBlame original commit: 1a2e88e23316d888e74ecce86de4eee362693a4b