Bug 758620 - Fix abort caused by unchecked clamping. r=ajuma
The initial values set for the offset for fixed position layers could clamp
with a max < min. This was due to a slightly incorrect min, but also because
it should have set these values in the else block of a later if that did a
range check on these values.

UltraBlame original commit: 8ca7a4f665a1ea213c51630a37a7f2483357161b