Bug 784878 - Use version 2 by default when no version in found in permissions.sqlite. r=sicking
This is a patch to workaround the fact that talos has wrongly gerenated
permissions.sqlite files which don't have a user_version and are version 2.
This is going to be replaced by a MOZ_ASSERT by bug 784884.

UltraBlame original commit: 9435261f7808c2d214e14ad8e97a255cee5a4ed6