Bug 786421. When hiding a popup panel don't bother sizing its view to zero because that is useless and confuses the widget sizing code. r=roc
Panels have a min size constraint set on them so when the view code tried to resize the widget to zero it would get bumped up to the min size. The view code would continue trying to resize the widget because of this.

UltraBlame original commit: e0ceffe737ddbbd7183403c7d96354ef04cd871d