Bug 783727 - Part 2: Define MACOSX_DEPLOYMENT_TARGET when populating virtualenv; r=glandium
Previously, we used a global version of MACOSX_DEPLOYMENT_TARGET which
was typically set to 10.6 because that is what Mozilla currently targets
as the minimum OS X version. Due to a bug in Python, using a
MACOSX_DEPLOYMENT_TARGET older than what Python itself was compiled with
results in an error. This patch is effectively a workaround for that
bug.

UltraBlame original commit: 980d552c34ac320c62514ff67bb2f84645f948a1