Bug 796839 - Part 4: Don't pretend that empty text nodes are not editable; r=roc
This is the real fix for this bug.  Previously we mistakenly thought
that if a text node is empty (or has only whitespace content), it is not
editable.  This patch removes that check completely and makes us treat
text nodes the same way that we treat element nodes.

UltraBlame original commit: e92287b62a547f0ae7cac5d8487e1159f4b13557