Bug 797213 - jstests.py: add a -F (--failed-only) option to only display output of failed tests. r=terrence
I also snuck a change to the output format, to display the test name in the output so that it is possible to associate output with the test that produced it. I mixed it in just because it was a little harder to separate out into a separate patch. I could easily do so, if you'd like.

UltraBlame original commit: 5a932c059384d73bb7e7d05c7386e9d352ecdd34