Bug 800983.  Expose a getter for default computed style.  r=dbaron
An alternative implementation strategy is to add "default-only"
versions of ResolveStyleFor and ResolvePseudoElementStyle and then
modify FileRules to support those, or use a cut-down custom version of
FileRules for them.  That would be faster, but more complicated on the
style set side.  We can always make that switch if we need to, I guess.

UltraBlame original commit: fd12a2c69e93b58cb79f6c3d0482d3df6943397d