Bug 800314 - Stop the browser window from stealing focus when a download begins. r=paolo.
The most recent browser window (if available) was stealing focus when a download
started. This was to account for the case where a download was initiated from
the Library. We are, however, switching to a Library view that shows download
state and progress in bug 675902, so this focus pull is now unnecessary, and
disruptive for users with window managers that don't have focus-stealing
prevention.

UltraBlame original commit: 83e6a9456aaf5611dc2a3a5021d3ab4b6cf82dbd