Bug 836301 - Stop enter()ing with CALL for nativeCall. r=mrbkap
This is kind of nonsensical, because CALL means "the wrapped object is being
called", whereas nativeCall means "the wrapped object is being unwrapped to
have a JSNative invoked on it", which are two very different things.

We _could_ add a NATIVECALL enter() trap, but our current policy enforcement
around nativeCall involves overriding the trap itself, so we wouldn't use it
for anything. So let's just get rid of it.

UltraBlame original commit: f976eaa7bbf4c2c64c9ffad46c7d2b3f70890341