Bug 850681 - Part 8: Only parse the leak log once; r=ted
The existing code parses the leak log twice - first time to output the
raw log lines & and the second to output the analysis of them.

As much fun as it is to loop all the things (\o/), let's not.

UltraBlame original commit: ae92e0819f2b0e925436aba44318dd571083cfcb