Bug 658909 - Remove GWNOJO from XPCVariant. r=mrbkap
The old code seems to be deciding whether we have a double-wrapped object by
checking _either_ the rv of GWNOJO _or_ the potential slim wrapper. This is
nonsensical, because double-wrapped objects are never slim wrappers.
Furthermore, that variable here is named 'proto', which further suggests
that this code is nonsensical. So let's just check for WNs.

Also, it seems pretty wack to be innerizing here before storing the jsval,
but I'm going to leave that for now.

UltraBlame original commit: d8ac57d4452f8e0e77dbd6deba4b9607200a3c7e