Bug 879341 - Move CheckForDebugMode into the main thread event observer. r=gabor
This gets invoked at the very end of nsThread::ProcessNextEvent(). If the cx stack
is empty here, that means the event loop isn't nested - if it it was, we'd at
least have a null cx on the stack from a higher-level call to OnProcessNextEvent
for the main thread observer.

As such, this seems like a much more reasonable place to make debug mode changes.

UltraBlame original commit: 4e9b95c25646d55fa9ae21e65f27b5c992e5ca84