Bumping gaia.json for 2 gaia-central revision(s)
========

https://hg.mozilla.org/integration/gaia-central/rev/67fc22e628e1
Author: Francisco Jordano <arcturusardeenelinfierno.com>
Desc: Merge pull request #10673 from wanderview/contacts-edit-noname

Bug 887564: Do not show phone # in name edit field

========

https://hg.mozilla.org/integration/gaia-central/rev/e4b3224b9a90
Author: Ben Kelly <benwanderview.com>
Desc: Bug 887564: Do not show phone # in name edit field

Previously the contacts_list code would overwrite the givenName field in
the contact object.  This was done so that different values could be
shown in the list if a name was not present.  This causes problems, though,
when the contact object is shared between multiple parts of the app.  In
this case the contacts_form code was getting handed one of these modified
contact objects.

Modify the code to calculate the display name, but not modify the contact
object itself.  Also refactor how contacts_list refresh() works when given
a contact object instead of an ID.  The previous code was a bit confusing
and called remove() twice for objects.

UltraBlame original commit: edda197b3f833933da4302ac114c31f5b35b8569