Bug 898811 - Use the system error reporter for XBL compilation. r=mrbkap
The current XBL error reporter logs things to the console service, but not to
stderr. And in certain situations (especially if there's a parse error in
frontend XBL that leaves the browser in a half-baked state), it can be very
difficult to get to the error console.

So we should log to stderr too, which is exactly what the system error reporter
does.

UltraBlame original commit: 57cda2d1472ff5aa8fee3cf71835084584dd00cd