Bug 903422: Replace AppendDeviceNameRunnable by non-blocking implementation, r=echou,gyeh
AppendDeviceNameRunnable currently block on the Bluetooth command thread
for a DBus reply.

This patch refactors the code to not block during the DBus call. The classes
AppendDeviceNameHandler and AppendDeviceNameRunnable are being removed, and
the DBus interaction is done directly in the DBus thread.

UltraBlame original commit: 13064db19b63ccc7139eef92f532f8a4726b51cb