Bug 941315 - Update the timing model even when invalid values are set; r=longsonr
When an invalid dur, min, max, or repeatDur is set we were failing to call
UpdateCurrentInterval to update the timing model. This patch makes sure we
update the current interval in error cases too.

Mochitests test the interval is being updated in these cases and, for
completeness, the case of repeatCount as well.

UltraBlame original commit: 00218997103be3b18b44b439533dcd19791385af