Bug 980256. nsDisplayZoom items need some adjustments if they are going to generate a scrollable layer. r=roc
Specifically GetBounds should use nsDisplaySubDocument::GetBounds (and not nsDisplayWrapList::GetBounds) before converting it to the right app units.

Similarly for ComputeVisibility except with the added wrinkle that nsDisplaySubDocument::ComputeVisibility won't properly handle coordinates if we are not generating a scroll layer. Namely it calls nsDisplayWrapList::ComputeVisibility which uses mVisibleRect (which is the visible rect of the scroll frame), when it needs to be using the visible rect of the contained list, which is for the scrolled content inside.

UltraBlame original commit: efbea5e8c766d808a6eb486d85fc404f4a12e67e