Bug 973976 - Revert to Thebes path for CopyableCanvasLayer. r=jmuizelaar
The performance on the Mac platform degraded after porting the code to
Moz2D in Bug 948765. This patch chooses the old thebes path instead of the
Moz2D path, so that performance is unaffected. This way we can easily
revert this patch at a later time when the perf issue has been fixed.

UltraBlame original commit: c737dc096ba0aaa6408686eb8fe9264fbaf10954