Bug 986367 - Make paint_listener.js check for paint suppression; r=mattwoodrow
This patch adds an extra check for paint suppression when waiting for paint
events. This is because on some platforms (notably B2G) we can think all paints
have completed because paint suppression is in effect and as a result call the
callback too soon.

This patch use window.setTimeout(..., 0) to wait for paint suppression to finish
before preceding to check for pending paint events.

When the refresh driver is under test control, if we detect that paint events
are pending we need to force a refresh driver tick. This patch adds that tick.
I suppose we had previously never hit this situation before and never noticed
this.

This patch also rearranges the main loop so that early returns appear first and
calling the callback appears at the end.

UltraBlame original commit: 827746f4c8ea981de02659df912c75b94d2e87f2