Bumping gaia.json for 2 gaia revision(s) a=gaia-bump
========

https://hg.mozilla.org/integration/gaia-central/rev/76e120554484
Author: Anthony Ricaud <anthonyricaud.me>
Desc: Merge pull request #20303 from Rik/keyboard-hide-toolbar-1019426

Bug 1019426 - When searching in the Address Book from the Dialer a white... r=etienne

========

https://hg.mozilla.org/integration/gaia-central/rev/646c8eb19edf
Author: Anthony Ricaud <anthonyricaud.me>
Desc: Bug 1019426 - When searching in the Address Book from the Dialer a white tab is shown over the keyboard

We previously had two ways to hide the toolbar so this is removing one for more consistency.
This also removes the transitions because they were weird and not working reliably.

UltraBlame original commit: 006411f793563a75b23ae57804d3dd011fcd1615