Bug 625289 patch 12 - Tests for transitions on elements that are reframing.  r=heycam
The second set of 14 tests all fail (both with and without the patch)
because we don't start transitions when the before state is
display:none (either for style contexts coming out of the undisplayed
map or because we have no current style context because an ancestor is
display:none).

The patch makes all of the first set of 14 tests pass.  Without the
patch, most of these tests fail, except for the 4 testing e1 and the
root.

I presume the four tests that pass prior to the patch pass because the
restyle actually occurs and starts the transition (which is stored on
the element) before the style context is destroyed.

UltraBlame original commit: 51037224ba032d9f1284c0f0dae6e0b576abefbc