Bug 1030896 - Import SwipeToDismissNOA. r=margaret SwipeToDismissNOA (https://github.com/JakeWharton/SwipeToDismissNOA) is a port of Roman Nurik's swipe-to-dismiss library to use NineOldAndroids. SwipeToDismissNOA adds swipe behavior to ListView that animates and removes the item.
As-is, there are a few problems with this library:
* The comments about the v12 API requirements are wrong
* It has problems with async adapters, such as CursorAdapter
* The MotionEvent#obtain calls are not followed by recycle
* There are beta warnings in the comments, likely due to the problems above

As a result, this library is being imported to mobile/android/base
instead of the normal thirdparty lib directory so we can make the
necessary modifications. Subsequent patches in bug 1030896 will aim to
address these issues to make this code production-ready.

UltraBlame original commit: 7fb7d61bc5d922c638c8ffcf4cc6335d5b4deec1