Bug 1097094 - Keep a WeakPtr to nsDocShell in SectionCarets. r=roc
When SelectionCarets::Terminate() is called, it's not guaranteed that we
can get nsDocShell from PresContext. It causes that SelectionCarets
cannot remove itself as an observer.

To fix this, we keep a member WeakPtr<nsDocShell> so that we can always
have nsDocShell in SelectionCarets::Terminate().

UltraBlame original commit: 12034172f9be3fc32afe220e60ef46f34ba2d339