Bug 1103998 - Add Flame bootloader to zip. r=nthomas
We do support building bootloader for the Flame, and this bootloader has
new features (limiting the number of CPUs, and maybe more in the
future). So people flashing from pvtbuilds should benefit from this to
ease their workflow, and not have to clone and build just to get the
bootloader.

UltraBlame original commit: 3f0f35b7204b1973534734f625b6a6d294093116