Bug 1107297: Only recomposite the damaged rect with D3D11. r=jrmuizel
This patch makes us behave similarly to when we're using the BasicCompositor, essentially we will clip all drawing to the final display to the area that was labeled as invalid. When DXGI 1.2 is available we will then also report the damaged area to the Present call so that can then be used to minimize the amount of bits that actually need to be blitted to the screen. Since we're no longer recompositing the whole screen this means we should also only clear the damaged area of the window.

UltraBlame original commit: 823227372483eb3b066c36df5f93036918560880