Bug 1109437 - Schedule the state machine thread in MediaSourceDecoder::Load. r=cpearce
This is necessary to get the state machine moving after it's initialized.
MediaDecoder::Load does this, but it looks like we missed this in the override.
So the current code relies on the ScheduleStateMachine call at the end of
TrackBuffer::AppendData to get things rolling. We're going to be removing that
call, so we need to fix this.

UltraBlame original commit: 4c38dda0688052e7074c04a7a25065cb7e3246c6