Bug 1121692 - Handle mid-seek Request{Audio,Video}Data calls. rpending=cpearce
The MDSM really shouldn't be issuing these. The correct solution (which I'll
file a followup bug for) is to fix the MDSM to not do that. But rejecting the
promise with CANCELED is a lot safer, so we do that for now.

UltraBlame original commit: 9bf200fcadbce728912d072eeb646b7c0291fcba