Bug 1122781 part 4 - Add runtime dependency check of style structs. r=dbaron
This patch adds a mechanism to check struct computation against a known set of non-cyclic dependencies.

The checks are only done when we get a struct on a context during computation of a struct on the same context.

UltraBlame original commit: 455e36d4b7ff5baeedfe32e9a8a9f5913ab34ae4