Bug 1119947 - Add telemetry for MSE video playback - r=cpearce,bsmedberg
This adds telemetry to record the state of the video playback
when the user exits. We are interested in knowing if the video
was buffering, paused, seeking, ended or other.

More telemetry will be added in bug 1127646.

UltraBlame original commit: fb653123d5552c4d4b3b9ce81d4f36738b1cce1c