Bug 1066812 - Set AbortOnError in the ContentProcess on initialization. r=billm.
This will cause the content process to take itself down in the event that it
loses communication with the parent process. This case is particularly
important for the case where the parent process crashes while the content
process is blocked or busy on the main thread, as the content process will
no longer continue to exist as a zombie process, but will shut down after
a short delay.

UltraBlame original commit: 09cc6c61c05941ed908f240edbd33731dbad7f13