Bug 1073462: Send synthetic property with Content:LocationChange message. r=felipe
The browser.isSynthetic property is needed by the zoom code to detect when to
apply the correct zoom level. In e10s it is currently only set when a new
document is created which means we don't set it right for history navigation.
This sends it with the Content:LocationChange message which is where it is
needed by the zoom code anyway.

UltraBlame original commit: 9f3479b12e5e27aae6a6286cc4ea9026498ee2cf