Bug 486172 - Don't ignore a -p command line argument when using the Xremote protocol. r=bsmedberg
On X11, when running firefox -p foo http://mozilla.org, and a window for another
profile is already open, the -p argument is ignored and a new tab or window is
opened in the unrelated session.

Previously, the equivalent firefox -p foo -remote openurl(http://mozilla.org)
would see that there is no window for the profile foo, complain about it, and
abort. If a window for the profile foo was open, however, a new tab or windows
would open in that session.

Here, we modify the behaviour such that firefox -p foo http://mozilla.org never
ignores the -p argument, and does the sensible thing depending on the context:
- if a window is already open for the profile, use that session.
- otherwise, open a new window for that profile.

When no -p argument is given, the behaviour is unchanged.

As RemoteCommandLine, which first attempts to open a connection with an existing
firefox, falls through when there is no existing firefox, the -p argument must be
kept in the command line. It turns out CheckArg didn't handle the case properly,
so fix this as well.

The changes in RemoteCommandLine otherwise match what used to be in
HandleRemoteArgument before bug 1080319.

UltraBlame original commit: f1902c4276e1f32ba569c571481d75751edff6b2