Bug 1148197 - Move share overlay margins to child to properly align. r=liuche
When it's on the parent of the outermost container with
layout_width="match_parent", the margin adds to the parent's size
(i.e. the screen width). Adding the margin to the child ensures the width is
reduced and content centered.

UltraBlame original commit: 521c3eb713749c37aeb0ec58646751f6893594c7