Bug 847287 patch 6 - Set mWinsInCascade for CSS Animations.  r=birtles
This is the main patch for the bug; it makes us use the mechanism added
in bug 1125455 to avoid sending animations that aren't currently
applying to the compositor.

Patch 7 is needed to make this code rerun in all the cases where we need
to rerun it, though.

UltraBlame original commit: de066f41c3776886288d112365b3fd57ffb882c0