Bug 1156456 patch 1 - Reftests comparing a 3-D transformed cube generated by main-thread paths and off-main-thread animations paths.  r=birtles
I confirmed that without the patch, the test marked as failing (the one
with zoom) in this patch (which will be unmarked in the next patch)
fails.

There is also a small amount of fuzz that will need to be annotated with
an additional patch on top of the patch stack.

UltraBlame original commit: dc6b401584e93fca684dee30f7c8b03c45caa78f