Bug 1159306 - explicitly initialize nsBaseHashtableET::mData; r=erahm
mData will get automatically constructed when DataType is a class or
struct with a default constructor, but a number of uses of
nsBaseHashtable use integer or pointer types as DataType.  Explicitly
initialize mData so that it looks as though we're fully initializing the
class in such cases.

UltraBlame original commit: f450fec85dd575a60d012afc2e5be65ba8753f19