Bug 1164767. Modify async scrollbar positioning code to deal with containerless root scrollables. r=botond
The root scrollbars will no longer be a descendant of the layer with their corresponding frame metrics. But scrollbars for root scrollables still need adjustment for the resolution.

We still use the descendant check for the code which needs to be run if the scrollbars are a descendant of the layer with their corresponding frame metrics so that non-containerless root scrolling still works the same way.

UltraBlame original commit: 1184021538f6ba44d79a18ff11eae5b5e6b6df2a