Bug 1168630. Part 4. Rename AsyncPanZoomController::IsRootForLayersId to HasNoParentWithSameLayersId. r=botond
The function did not tell you if the APZC was the root for the layers id. It just told you if it had no parent. Which are different things.

Since IsRootForLayersId didn't do what it was expected to, the users of HasNoParentWithSameLayersId will be audited and then removed or changed to be correct by bug 1158424.

UltraBlame original commit: 44a885179fc3c484c730cdc2d6564cb6d574abd8