Bug 1160695 - Fix up test to require the true duration, rather than the fake duration. r=jww,sr=cpearce
From the bug it corresponds to, this test was designed to make sure that we don't
read ogg duration by seeking (which causes network activity) when the server
offers the X-Content-Duration header. However, preserving this behavior requires
the ogg reader to be aware of the presence of network duration when it's reading
metadata, which isn't easily accomplished in our new model here. I think both ogg
and X-Content-Duration are sufficiently unimportant at this point that this is an
acceptable regression.

UltraBlame original commit: f80895a195ef99fd3de3eced9f9754a4f0d35cb2