Bug 1163223 - Switch test_buffered to use timeupdate rather than loadedmetadata. r=cpearce
The problem is that we can't compute buffered until we know the start time, but don't
necessarily know the start time by the time we fire loadedmetadata (we just know the
duration). Chris determined that there was nothing in the spec that requires any
particular behavior from .buffered after loadedmetadata, so this should be fine.

UltraBlame original commit: 7b6804398fc3aeeea8b61146a88748ce432afc50