Bug 1170864 - Refactor DevToolsUtils.fetch to use NetUtil more extensively and make it fall back to OS.File if a local file cannot be read. r=past
The method was using NetUtil or IOService based on the scheme of the given
URL. Due to changes in NetUtil since the code was originally written these two
code paths are doing the same thing now. These changes unify the two cases into
a single one which
* creates a channel for the given URL and works around some problems in
  xpcshell-tests
* sets loading options on the channel and
* leaves the fetching, stream reading and charset conversion to NetUtil.

It also adds code to work around bug 982654 which causes the stream to
throw if the file is empty. If a file cannot be read by the file:// handler, it
will try to use OS.File.read which handles empty files properly. This is only
used as a fallback as OS.File doesn't provide the content type for the file
required by the method.

UltraBlame original commit: a6d0ac7916ff5017d9c0bd7cf7e43098fba8f2d7