Bug 1190669 part 1 - Remove helper function SetWindowFullScreen and its helper runnable. r=smaug
All of the callsites of this helper function has a synchronous event
dispatch which indicates the AddScriptRunner in that helper function
makes no sense.

Also it seems that most of the callsites are actually safe to run
script, except ExitFullscreenInDocTree() which may be called from
Element::UnbindFromTree(). It'll be fixed in the following patch.

UltraBlame original commit: 2d09363a1516e74bcab7514643f8b5c77b91e65a