Bug 1196970 - Part 6: Copy preprocessed resources from objdir to build directory during Gradle build. r=sebastian DONTBUILD NPOTB
This means we don't require the directory in the object directory at
gradle-install time.  We're not concerned if the resource files are
missing, since we have code to ensure they're fresh already; and if
they are missing, we'll quickly fail as we try to process the resource
set.

UltraBlame original commit: c92b5eb419b09ca81fcde189486435a6ce7f41f4