Bug 1197185 - Give NonE10s a virtual destructor; r=jesup
The Parent template class inherits from both NonE10s and PMediaParent,
and the latter has a virtual destructor.  This confuses clang-tidy when
it wants to rewrite ~Parent() which is currently marked as virtual, and
it makes ~Parent() override which breaks the build since ~NonE10s() is
not virtual.  The easiest way to work around this seems to be making
~NonE10s() virtual.  See the discussion in
https://llvm.org/bugs/show_bug.cgi?id=24496 as well.

UltraBlame original commit: ee7879fae95371b6dfbf83baa537ac53d9cddfd3