Bug 1195517 - Remove unused ic_menu_addons_filler. r=liuche
This also removes code to set an icon from an addon, which is unused because,
afaik, addons' icons are also hidden. iirc, there was a bug open on whether we
want addons to be able to show icons (e.g. could be used to show status), but
we can reimplement this if that bug is ever decided.

I was not able to test this with an addon that sets an icon but I did test that
the application did not crash with no addons installed.

UltraBlame original commit: 27be4ca1e6c9602b9c6aa718700e137e3f4f8361